Refactor TransactionCreate
